requirements.txt
# Core Django & API
Django>=5.1
djangorestframework>=3.15.0
django-cors-headers>=4.3.1

# Async & WebSockets
channels>=4.1.0
channels-redis>=4.2.0
daphne>=4.1.0

# Database
psycopg[binary]>=3.2.0
dj-database-url>=2.1.0

# Celery & Redis
celery>=5.4.0
django-celery-beat>=2.6.0
django-celery-results>=2.5.1
redis>=5.0.1
django-redis>=5.4.0

# Networking & Security
requests>=2.32.0
urllib3>=2.2.0
cryptography>=42.0.0
pyOpenSSL>=24.0.0
service-identity>=24.1.0
twisted[tls]>=24.3.0
httpx>=0.28.1

# Utilities
python-dotenv>=1.0.1
async-timeout>=4.0.3
attrs>=23.2.0
cffi>=1.16.0
pytz>=2024.1
sqlparse>=0.5.0
typing_extensions>=4.10.0

# Production Server
gunicorn>=22.0.0
uvicorn>=0.30.0
websockets>=12.0
whitenoise>=6.6.0


# Object storage (Cloudflare R2 / S3)
django-storages[boto3]>=1.14.0
boto3>=1.34.0

# AI/ML Integration
huggingface-hub>=0.26.0
anthropic>=0.40.0
openai>=1.50.0

# Authentication & Social
django-allauth>=65.0.0
PyJWT>=2.8.0
requests-oauthlib>=1.3.1

# Third-party Integrations
temporalio>=1.2.0
amadeus>=8.0.0
intasend-python>=0.1.0
pydub>=0.25.1
twilio>=9.0.0

# Data handling
Pillow>=10.0.0
pypdf>=5.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Security hardening
django-axes>=6.0.0
django-ratelimit>=4.1.0
django-csp>=3.8.0
django-auditlog>=3.0.0
bleach>=6.1.0
pycryptodome>=3.17.0
python-json-logger>=2.0.0

# Development & Testing (Optional)
# Install with: pip install -r requirements.txt[dev]
# Uncomment the following lines locally only:
# pytest>=7.4.0
# pytest-django>=4.5.0
# pytest-cov>=4.1.0
# factory-boy>=3.2.0
# black>=23.0.0
# flake8>=6.0.0
# isort>=5.12.0
# mypy>=1.0.0
# django-stubs>=1.14.0
# django-debug-toolbar>=3.8.0
# django-extensions>=3.2.0
# Sphinx>=6.0.0
# sphinx-rtd-theme>=1.2.0
# safety>=2.3.0
# bandit>=1.7.4